[KYUUBI #3893] [BUG] Fix spark extension: UnspecifiedDistribution does not have default partitioning.
